chore: refactor markdown file, pull out completions
